abstract comment form + scroll into view automatically
