+++ frontend/src/components/CodeBrowser/CodeLine/styles.css
@@ -67,181 +67,4 @@
     margin: 0.75em 0;
     font-size: 13px;
   }
-
-  &__comment {
-    padding: 12px 20px;
-    width: 100%;
-    border: 1px solid var(--comment-border);
-    background: var(--comment-background);
-    border-radius: 4px;
-    font-family: var(--primary-font);
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    color: #a3a3a3;
-
-    &Head {
-      display: flex;
-      flex-direction: row;
-      gap: 10px;
-      align-items: center;
-      font-weight: 500;
-      color: #dfdfdf;
-      font-size: 14px;
-
-      img {
-        width: 26px;
-        height: 26px;
-        border-radius: 100%;
-      }
-    }
-
-    &Body {
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      gap: 14px;
-    }
-
-    &Pending {
-      --pending-color: #ffaf2e;
-      display: flex;
-      width: fit-content;
-      flex-direction: row;
-      color: var(--pending-color);
-      padding: 1px 14px;
-      border-radius: 100vw;
-      font-weight: 400;
-      border: 2px solid var(--pending-color);
-      font-size: 11px;
-    }
-
-    &Points {
-      display: flex;
-      width: fit-content;
-      flex-direction: row;
-      gap: 5px;
-      color: white;
-      padding: 3px 8px;
-      border-radius: 3px;
-      font-weight: 500;
-
-      &--positive {
-        background: var(--green);
-      }
-      &--negative {
-        background: var(--red);
-      }
-      &--neutral {
-        background: var(--comment-border);
-      }
-    }
-  }
-
-  &__newCommentForm {
-    display: flex;
-    width: 100%;
-    flex-direction: column;
-    gap: 10px;
-
-    textarea,
-    input {
-      resize: vertical;
-      width: 100%;
-      outline: none;
-      background: rgba(0, 0, 0, 0.1);
-      border: 1px solid var(--comment-border);
-      border-radius: 4px;
-      padding: 10px;
-      color: white;
-      font-family: var(--primary-font);
-
-      &:focus {
-        border-color: var(--brand-lighter);
-      }
-    }
-
-    .CodeLine__newCommentPoints {
-      display: flex;
-      align-items: center;
-      position: relative;
-      height: 40px;
-
-      label {
-        margin-right: 10px;
-      }
-      input {
-        height: 100%;
-        width: 60px;
-        text-align: center;
-        -webkit-appearance: textfield;
-        -moz-appearance: textfield;
-        appearance: textfield;
-        border-right: none;
-        border-bottom-right-radius: 0;
-        border-top-right-radius: 0;
-
-        &::-webkit-inner-spin-button,
-        &::-webkit-outer-spin-button {
-          -webkit-appearance: none;
-        }
-
-        &:focus ~ .CodeLine__newCommentPoints__spinners {
-          border-color: var(--brand-lighter);
-        }
-      }
-
-      &__spinners {
-        height: 100%;
-        border: 1px solid var(--comment-border);
-        border-left: none;
-        border-bottom-right-radius: 4px;
-        border-top-right-radius: 4px;
-
-        > div {
-          background: rgba(0, 0, 0, 0.1);
-          height: 50%;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          width: 1.5em;
-          cursor: pointer;
-          user-select: none;
-          border-left: 1px solid var(--comment-border);
-
-          &:first-of-type {
-            border-bottom: 1px solid var(--comment-border);
-            &:active {
-              background: var(--green);
-            }
-          }
-          &:last-of-type {
-            &:active {
-              background: var(--red);
-            }
-          }
-
-          &:hover {
-            background: var(--comment-border);
-          }
-        }
-      }
-    }
-
-    .CodeLine__newCommentButtons {
-      display: inline-flex;
-      gap: 5px;
-
-      > .Button {
-        flex-shrink: 0;
-        padding: 8px 16px;
-        font-weight: 500;
-      }
-
-      .CodeLine__newCommentCancel {
-        background: var(--comment-border);
-        border: 1px solid var(--comment-border);
-      }
-    }
-  }
 }
